requirements.txt
pyyaml
paho-mqtt
urllib